rename playbackrange to timeline
